+++ test/fixtures/src/inflater_spec.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+// Define constants for all widgets and objects we want to generate inflaters for.
+// Used by the builder to generate inflaters.
+
+// ignore_for_file: unused_element
+
+const AppBar? _appBar = null;
+const Center? _center = null;
+const Column? _column = null;
+const FloatingActionButton? _floatingActionButton = null;
+const Icon? _icon = null;
+const MaterialApp? _materialApp = null;
+const Scaffold? _scaffold = null;
+const Text? _text = null;
+const TextStyle? _textStyle = null;
+const ThemeData? _themeData = null;